Rename foundation package to common per review feedback and update lint config; remove duplicated logic in trace xrm client that already made it to main in another PR
