+++ .golangci.yml
@@ -220,6 +220,12 @@ linters:
           - staticcheck
         text: 'SA1019: .+ is deprecated: Use Composition Functions instead.'
 
+      # We include a 'common' package under crank, which revive doesn't like.
+      - linters:
+          - revive
+        path: 'cmd/crank/common/*'
+        text: 'var-naming: avoid meaningless package names'
+
     paths:
       - zz_generated\..+\.go$
       - .+\.pb.go$
