+++ cmd/crank/beta/trace/internal/printer/default.go
@@ -31,8 +31,8 @@ import (
 	"github.com/crossplane/crossplane-runtime/v2/pkg/fieldpath"
 
 	pkgv1 "github.com/crossplane/crossplane/v2/apis/pkg/v1"
-	"github.com/crossplane/crossplane/v2/cmd/crank/beta/trace/internal/resource"
-	"github.com/crossplane/crossplane/v2/cmd/crank/beta/trace/internal/resource/xpkg"
+	"github.com/crossplane/crossplane/v2/cmd/crank/common/resource"
+	"github.com/crossplane/crossplane/v2/cmd/crank/common/resource/xpkg"
 	"github.com/crossplane/crossplane/v2/internal/xcrd"
 )
