+++ cmd/crank/beta/trace/internal/printer/default_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/v2/pkg/test"
 
-	"github.com/crossplane/crossplane/v2/cmd/crank/foundation/resource"
+	"github.com/crossplane/crossplane/v2/cmd/crank/common/resource"
 )
 
 func TestDefaultPrinter(t *testing.T) {
