+++ cmd/crank/beta/trace/internal/printer/dot.go
@@ -12,8 +12,8 @@ import (
 	"github.com/crossplane/crossplane-runtime/v2/pkg/fieldpath"
 
 	v1 "github.com/crossplane/crossplane/v2/apis/pkg/v1"
-	"github.com/crossplane/crossplane/v2/cmd/crank/foundation/resource"
-	"github.com/crossplane/crossplane/v2/cmd/crank/foundation/resource/xpkg"
+	"github.com/crossplane/crossplane/v2/cmd/crank/common/resource"
+	"github.com/crossplane/crossplane/v2/cmd/crank/common/resource/xpkg"
 )
 
 // DotPrinter defines the DotPrinter configuration.
