+++ cmd/crank/beta/trace/internal/printer/dot_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/v2/pkg/test"
 
-	"github.com/crossplane/crossplane/v2/cmd/crank/foundation/resource"
+	"github.com/crossplane/crossplane/v2/cmd/crank/common/resource"
 )
 
 // Define a test for PrintDotGraph.
