+++ cmd/crank/beta/validate/cmd.go
@@ -30,7 +30,7 @@ import (
 	"github.com/crossplane/crossplane-runtime/v2/pkg/errors"
 	"github.com/crossplane/crossplane-runtime/v2/pkg/logging"
 
-	"github.com/crossplane/crossplane/v2/cmd/crank/foundation/load"
+	"github.com/crossplane/crossplane/v2/cmd/crank/common/load"
 	"github.com/crossplane/crossplane/v2/internal/version"
 )
