+++ cmd/crank/common/load/testdata/folder/nested-folder/resource-a.yaml
