+++ cmd/crank/common/load/testdata/folder/resource-b.yaml
