+++ cmd/crank/common/package.go
@@ -1,0 +1,2 @@
+// Package common provides functions and helpers shared between CLI utilities, available for import by external tools interacting with Crossplane.
+package common
