+++ cmd/crank/common/resource/xpkg/client_test.go
@@ -20,7 +20,7 @@ import (
 
 	xpkgv1 "github.com/crossplane/crossplane/v2/apis/pkg/v1"
 	"github.com/crossplane/crossplane/v2/apis/pkg/v1beta1"
-	"github.com/crossplane/crossplane/v2/cmd/crank/foundation/resource"
+	"github.com/crossplane/crossplane/v2/cmd/crank/common/resource"
 )
 
 // TODO add more cases, fake client
